+++ docs/bundle.js
@@ -77,6 +77,8 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 
+var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
+
 var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }(); /**
                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       * textoverlay.js - Simple decorator for textarea elements
                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       *
@@ -139,7 +141,7 @@ var Textoverlay = function () {
       var position = (0, _getStyle2.default)(textarea, ["position"]).position;
       var wrapper = document.createElement("div");
       wrapper.className = "textoverlay-wrapper";
-      (0, _setStyle2.default)(wrapper, Object.assign({}, (0, _getStyle2.default)(textarea, properties.wrapper), css.wrapper, {
+      (0, _setStyle2.default)(wrapper, _extends({}, (0, _getStyle2.default)(textarea, properties.wrapper), css.wrapper, {
         position: position === "static" ? "relative" : position
       }));
       parentElement.insertBefore(wrapper, textarea);
@@ -153,7 +155,7 @@ var Textoverlay = function () {
       var overlay = document.createElement("div");
       overlay.className = "textoverlay";
 
-      (0, _setStyle2.default)(overlay, Object.assign({}, css.overlay, (0, _getStyle2.default)(textarea, properties.overlay)));
+      (0, _setStyle2.default)(overlay, _extends({}, css.overlay, (0, _getStyle2.default)(textarea, properties.overlay)));
       wrapper.insertBefore(overlay, textarea);
       return overlay;
     }